eval/requirements.txt
ragas>=0.2.0
pandas>=2.0.0
pyarrow>=14.0.0
datasets>=2.16.0
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-google-genai>=1.0.0
litellm>=1.40.0
